+++ demo/src/main/java/org/vaadin/stefan/ui/view/demos/customtimeline/FixedDaysCalendarView.java
@@ -1,12 +1,21 @@
 package org.vaadin.stefan.ui.view.demos.customtimeline;
 
+import elemental.json.Json;
+import elemental.json.JsonObject;
 import org.vaadin.stefan.fullcalendar.CalendarView;
 
 import elemental.json.JsonFactory;
 import elemental.json.impl.JreJsonFactory;
 import elemental.json.impl.JreJsonObject;
+import org.vaadin.stefan.fullcalendar.CustomCalendarView;
+import org.vaadin.stefan.ui.view.demos.HasIntervalLabel;
 
-public class FixedDaysCalendarView implements CalendarView {
+import java.text.DateFormat;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.util.Locale;
+
+public class FixedDaysCalendarView implements CustomCalendarView, HasIntervalLabel {
 	private final int numberOfDays;
 
     public FixedDaysCalendarView(int numberOfDays) {
@@ -15,40 +24,28 @@ public class FixedDaysCalendarView implements CalendarView {
 
     @Override
     public String getClientSideValue() {
-        return "customView";
-    }
-
-    /**
-     * views: {
-     * customView: {
-     * type: 'timeline',
-     * duration: { days: 31 }
-     * }
-     * }
-     *
-     * @return
-     */
-    public JreJsonObject getInitialOptions() {
-        JsonFactory factory = new JreJsonFactory();
-        JreJsonObject initialOptions = new JreJsonObject(factory);
-
-        JreJsonObject durationHolder = new JreJsonObject(factory);
-        durationHolder.set("days", factory.create(numberOfDays));
-
-        JreJsonObject customViewHolder = new JreJsonObject(factory);
-        customViewHolder.set("type", factory.create("resourceTimeline"));
-        customViewHolder.set("duration", durationHolder);
-
-        JreJsonObject viewsHolder = new JreJsonObject(factory);
-        viewsHolder.set(getName(), customViewHolder);
-
-        initialOptions.set("views", viewsHolder);
-
-        return initialOptions;
+        return "fixedDaysResourceTimeline";
     }
 
     @Override
     public String getName() {
-        return "customView";
+        return "Fixed Days Resource Timeline";
+    }
+
+    @Override
+    public JsonObject getViewSettings() {
+        JsonObject days = Json.createObject();
+        days.put("days", numberOfDays);
+
+        JsonObject baseSettings = Json.createObject();
+        baseSettings.put("type", "resourceTimeline");
+        baseSettings.put("duration", days);
+        return baseSettings;
+    }
+
+    @Override
+    public String formatIntervalLabel(LocalDate intervalStart, Locale locale) {
+        DateTimeFormatter pattern = DateTimeFormatter.ofPattern("dd.MM.yyyy");
+        return pattern.format(intervalStart) + " - " + pattern.format(intervalStart.plusDays(numberOfDays - 1));
     }
 }